Refactor to move my code into it's own module.
