+++ package.json
@@ -24,6 +24,7 @@
     "@ng-bootstrap/ng-bootstrap": "^1.0.4",
     "bootstrap": "^4.0.0",
     "core-js": "^2.4.1",
+    "eslint": "^4.19.1",
     "rxjs": "^5.5.6",
     "zone.js": "^0.8.19"
   },
