+++ src/app/stratton/components/stratton.component.ts
@@ -1,0 +1,18 @@
+import { Component, OnInit } from '@angular/core';
+
+  @Component({
+    selector: 'app-stratton',
+    templateUrl: '../templates/stratton.component.html',
+    styleUrls: ['../content/stratton.component.css']
+  })
+  export class StrattonComponent implements OnInit {
+
+    constructor() { }
+
+    ngOnInit() {
+
+    }
+
+  }
+
+
