--- src/app/stratton/stratton.component.html
@@ -1,3 +1,0 @@
-<p>
-  stratton works!
-</p>
