--- src/app/stratton/stratton.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-stratton',
-  templateUrl: './stratton.component.html',
-  styleUrls: ['./stratton.component.css']
-})
-export class StrattonComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
