+++ src/app/stratton/templates/stratton.component.html
@@ -1,0 +1,3 @@
+<p>
+  stratton works! see
+</p>
